Move stuff to get a clean layout file
